+++ satella/coding/algos.py
@@ -4,6 +4,11 @@ __all__ = ['merge_dicts']
 
 
 def merge_dicts(v1: tp.Any, v2: tp.Any) -> tp.Any:
+    """
+    Try to merge two dicts/list together. If key collision is found, value from v2 will be taken.
+
+    Lists will be concatenated, and dicts updated. v1 will be updated in-place!
+    """
     if isinstance(v1, dict) and isinstance(v2, dict):
         for k in v2.keys():
             try: