+++ satella/coding/fun_static.py
@@ -1,28 +1,21 @@
-import logging
-import typing
-import functools
-
-logger = logging.getLogger(__name__)
-
-
-def static_var(var_name, value):
+def static_var(var_name: str, starting_value):
     """
     Declare a static variable for given function
 
     Use it like:
 
-    @static_var('counter', 2)
-    def count():
-        count.counter += 1
+    >>> @static_var('counter', 2)
+    >>> def count():
+    >>>     count.counter += 1
 
     or:
 
-    class MyClass:
-        @static_var('counter', 2)
-        def count():
-            MyClass.counter += 1
+    >>> class MyClass:
+    >>>     @static_var('counter', 2)
+    >>>     def count():
+    >>>         MyClass.counter += 1
     """
     def decorate(func):
-        setattr(func, var_name, value)
+        setattr(func, var_name, starting_value)
         return func
     return decorate